Fixed delete[] -> delete on vector<string>*, more error codes, check for ELF signature

+++ elfs/terminal/main.cpp
@@ -14,7 +14,7 @@
 extern "C" int memcmp ( const void * ptr1, const void * ptr2, size_t num ) {
     const char *p1 = (const char*)ptr1;
     const char *p2 = (const char*)ptr2;
-    for (int i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
         if (p1[i] < p2[i])
             return -1;
         else if (p1[i] > p2[i])
